+++ app/screens/addEpenses.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Text, View } from 'react-native'
+
+const AddEpenses = () => {
+    return (
+        <View>
+            <Text>
+                Expenses will be add here
+            </Text>
+        </View>
+    )
+}
+
+export default AddEpenses